remove extra boxing, redundant public declarations, and redundant final finals

+++ nars_gui/src/main/java/automenta/vivisect/swing/property/sheet/editor/PercentageEditor.java
@@ -58,35 +58,35 @@ public class PercentageEditor extends SpinnerEditor {
 		Number step = null;
 
 		if (type == Byte.class || type == byte.class) {
-			start = Byte.valueOf((byte) pstart);
-			min = Byte.valueOf((byte) pmin);
-			max = Byte.valueOf((byte) pmax);
-			step = Byte.valueOf((byte) pstep);
+			start = (byte) pstart;
+			min = (byte) pmin;
+			max = (byte) pmax;
+			step = (byte) pstep;
 		} else if (type == Short.class || type == short.class) {
-			start = Short.valueOf((short) pstart);
-			min = Short.valueOf((short) pmin);
-			max = Short.valueOf((short) pmax);
-			step = Short.valueOf((short) pstep);
+			start = (short) pstart;
+			min = (short) pmin;
+			max = (short) pmax;
+			step = (short) pstep;
 		} else if (type == Integer.class || type == int.class) {
-			start = Integer.valueOf(pstart);
-			min = Integer.valueOf(pmin);
-			max = Integer.valueOf(pmax);
-			step = Integer.valueOf(pstep);
+			start = pstart;
+			min = pmin;
+			max = pmax;
+			step = pstep;
 		} else if (type == Long.class || type == long.class) {
-			start = Long.valueOf(pstart);
-			min = Long.valueOf(pmin);
-			max = Long.valueOf(pmax);
-			step = Long.valueOf(pstep);
+			start = (long) pstart;
+			min = (long) pmin;
+			max = (long) pmax;
+			step = (long) pstep;
 		} else if (type == Float.class || type == float.class) {
-			start = Float.valueOf(pstart);
-			min = Float.valueOf(pmin);
-			max = Float.valueOf(pmax);
-			step = Float.valueOf(pstep);
+			start = (float) pstart;
+			min = (float) pmin;
+			max = (float) pmax;
+			step = (float) pstep;
 		} else if (type == Double.class || type == double.class) {
-			start = Double.valueOf(pstart);
-			min = Double.valueOf(pmin);
-			max = Double.valueOf(pmax);
-			step = Double.valueOf(pstep);
+			start = (double) pstart;
+			min = (double) pmin;
+			max = (double) pmax;
+			step = (double) pstep;
 		} else if (type == BigDecimal.class) {
 			start = new BigDecimal(pstart);
 			min = new BigDecimal(pmin);
